Testing the refactored features
