+++ src/test/java/test/dashboard/HomeControllerTest.java
@@ -1,4 +1,4 @@
-package test.controller;
+package test.dashboard;
 
 import static org.junit.Assert.assertEquals;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
